Minor code reorg
