+++ erftools/preprocessing.py
@@ -1,12 +1,15 @@
 import numpy as np
+import xarray as xr
 import f90nml
 
 from .wrf_namelist_input import time_control, domains, physics
-from .inputs import ERFinput
+from .inputs import ERFInputFile
 
-class WRFnamelist(object):
-    """Class to parse inputs from a WRF namelist.input file and convert to
-    inputs for ERF
+class WRFInputDeck(object):
+    """Class to parse inputs from WRF and convert to inputs for ERF
+    WRF inputs include:
+    * namelist.input
+    * wrfinput_d01[, wrfinput_d02, ...]
     """
 
     def __init__(self,nmlpath):
@@ -15,7 +18,7 @@ class WRFnamelist(object):
         self.time_control = time_control(self.nml['time_control'])
         self.domains = domains(self.nml['domains'])
         self.physics = physics(self.nml['physics'])
-        self.erf_input = ERFinput()
+        self.erf_input = ERFInputFile()
         self.calculate_inputs()
 
     def __str__(self):
@@ -25,14 +28,21 @@ class WRFnamelist(object):
         return s
 
     def calculate_inputs(self):
+        """Scrape inputs for ERF from a WRF namelist.input file
+
+        Notes:
+        * This does not import the WRF height levels (geopotential height); to
+          get those coordinates, call get_heights() to extract from WRF initial
+          conditions (`wrfinput_d01` from real.exe)
+        * This does not import z0 (surface roughness) from the reanalysis data;
+          to get that, call get_roughness_map()
+        """
         tdelta = self.time_control.end_datetime - self.time_control.start_datetime
         self.erf_input['stop_time'] = tdelta.total_seconds()
 
         # note: number vert pts is staggered
         # TODO: add vertical stretching
         n_cell = np.array([self.domains.e_we[0], self.domains.e_sn[0], self.domains.e_vert[0]-1])
-        #print('destaggered (cell center) heights:',heights)
-        #heights = self.domains.get_heights()
         ztop_est = 287.0 * 300.0 / 9.81 * np.log(1e5/self.domains.p_top_requested)
         self.erf_input['geometry.prob_extent'] = n_cell * np.array([self.domains.dx[0], self.domains.dy[0], np.nan])
         self.erf_input['geometry.prob_extent'][2] = ztop_est
@@ -61,4 +71,23 @@ class WRFnamelist(object):
         # TODO: specify PBL scheme per level
         self.erf_input['erf.pbl_type'] = self.physics.bl_pbl_physics[0]
         
+        return self.erf_input
         
+    def get_heights(self):
+        # Note: This calculates the heights for the outer WRF domain only
+        wrfinp = xr.open_dataset('wrfinput_d01')
+        ph = wrfinp['PH'] # perturbation geopotential
+        phb = wrfinp['PHB'] # base-state geopotential
+        hgt = wrfinp['HGT'] # terrain height
+        self.terrain = hgt
+        geo = ph + phb # geopotential, dims=(Time: 1, bottom_top_stag, south_north, west_east)
+        geo = geo/9.81 - hgt
+        geo = geo.isel(Time=0).mean(['south_north','west_east']).values
+        self.heights = (geo[1:] + geo[:-1]) / 2 # destaggered
+        # TODO: update erf_input with specified height levels
+        return self.heights
+
+    def get_roughness_map(self):
+        wrfinp = xr.open_dataset('wrfinput_d01')
+        
+
